Added script to build dependencies
This project requires presence of libusb, libftdi and libdivecomputer libraries
on the system. Hence, these are cross compiled for usage on android. The
generated static libraries can then be used for JNI interfacing.
This script assumes presence of libdivecomputer, libusb and libftdi submodules
in crossbuild folder.
To make, run the command as per target arch : bash build.sh [x86 / arm]

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>