View initialisation and scratching JNI
We are starting to scratch the surface of JNI. In this commit, we have taken all
the required views and initialised them. Also, a class contaning all native
functions is included. Right now, the aim it to populate the product and vendor
spinners by getting a list from libdivecomputer library.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>