Building native parts for JNI
For making a function that gives back the String array, we are trying to use the
libdivecomputer library. These functions are meant to fill up the product and
vendor spinners on the application. NativeHelper is incidently a module in
bionic library and was cause of no JNI_OnLoad method errors. Hence it is removed
and the functions migrated to Main.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>