Ask for permission if not already have.
In case we do not have permission for usage of USB device, ask the user for
permission. This happens on clicking OK button. On giving permmission, take us
to ImportProgress activity.
Also changed some wrapping policies. Because the line generally expand to much
larger dimensions than 80 and auto wrap seems to be making it more ugly, I think
its prudent not to wrap.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>