TextView to input filename
The earlier textviews are now changed to edittext to help user enter the logfile
and dumpfile names. They are enabled disables as per the checkbox.
The input file name is sent to the backend for opening the logfile. Native
methods for setting logfile and empty method for import of dives is added.
Some changes appearing in the commit is due to xml autoformat of eclipse.
Some other changes include setting of unregisterReciever of the broadcast on
destruction of activity.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>