Check for Storage and send location for logfile
On clicking OK, check if external storage is in writable condition. This check
earlier was happening in ImportProgress. I think this is more suitable here.
This way, ImportProgress has only one task, import the xml files and show user
the progress.
Name the default log and bin file better. Also now, the complete path is
transferred to the native side.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>