Massive changes - No more subsurface
This commit contains huge changes in the approach for this application. The
previous approach was to use subsurface directly to do all the necessary tasks
of downloading dives and doing all kinds of stuff with it. But due to the
inherent complexity of that task and lack of time, I am switching to a much
simpler approach of only using libdivecomputer for importation. This makes sense
as the application is aimed at only achieving import not being a subsurface
application. This was mentioned by my mentor but due to my ignorance, I could
not understand it at that time.

Now, the import task is heavily based on the universal script of
libdivecomputer. More accurately, util, common and universal are taken directly
from libdivecomputer/examples directory with minute changes.

JNI wrappers are more sensible now with well defined tasks.

Subsurface submodule is not removed yet. (Will be removed if this branch is
merged with master.)

As subsurface is removed, the script name has been changed from subsurface_jni
to libdivecomputer_jni. Appropriate changes are made in Android.mk.
( It does not build libdivecomputer_jni yet. )

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>