More OOP based changes in java side.
Making the system simpler and more robust, and more Object Oriented, I have
added all the native methods to DcData class except the getDeviceMap method.
Also with the added exception class, it will be easier to notify user of what is
actually going on.
Rather than showing a new activity with import progress, I am planning on
showing a dialog or a progress bar directly.

Main class has not been changed yet to incorporate the new architecture of
application.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>