Checks in DcData for proper input
DcData should contain values in particular form.
For ex. if vendor or product name is empty, its an error. Similarly if outfile
or dumpfile is empty, and it is needed, its an error. Similar is the case of non
positive file descriptor for opened usb device.
Added checks for these. Any of these conditions is attempted will result in
exception being thrown.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>