Initialise descriptor and fix lots of errors.
Initialise dc_descriptor before trying to import. The dive import was failing
due to this. Also added other steps for import of dives.
The changes made in libdivecomputer branch had to be incorporated in subsurface
branch. But some of the stuffs were left. These were causing errors such as
stacked dialogboxes and import failures. These were taken care of in this
commit.
Also, due to eclipse auto-formatter, the commit seems much larger than it
actually is.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>